frontend refactoring, changed project structure
