+++ employee-manager-app/src/app/app.module.ts
@@ -5,9 +5,15 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { EmployeeComponent } from './employee/employee.component';
-import { LoginComponent } from './login/login.component';
-import { UserComponent } from './user/user.component';
+import { EmployeeComponent } from './components/employee/employee.component';
+import { LoginComponent } from './components/login/login.component';
+import { UserComponent } from './components/user-employee/user.component';
+import { NavbarComponent } from './components/header/header-navbar.component'
+import { InfoModalComponent } from './components/employee/info-employee/info-modal.component';
+import { UpdateModalComponent } from './components/employee/update-employee/update-modal.component';
+import { AddModalComponent } from './components/employee/add-employee/add-modal.component';
+import { CardComponent } from './components/employee/card-employee/card-employee.component';
+import { DeleteModalComponent } from './components/employee/delete-employee/delete-modal.component';
 
 
 @NgModule({
@@ -15,7 +21,13 @@ import { UserComponent } from './user/user.component';
     AppComponent,
     EmployeeComponent,
     LoginComponent,
-    UserComponent
+    UserComponent,
+    NavbarComponent,
+    InfoModalComponent,
+    UpdateModalComponent,
+    AddModalComponent,
+    CardComponent,
+    DeleteModalComponent
   ],
   imports: [
     BrowserModule,
