+++ employee-manager-app/src/app/components/employee/employee.component.html
@@ -1,0 +1,4 @@
+<header-navbar></header-navbar>
+<card [infoEmployee]="infoEmployee" [deleteEmployee]="deleteEmployee" [editEmployee]="editEmployee" [infoEmployee]="infoEmployee" ></card>
+
+
