+++ employee-manager-app/src/app/components/header/header-navbar.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core'
+import { ModalService } from 'src/app/services/modal.service'
+import { Employee } from '../employee/employee'
+
+@Component({ templateUrl: 'header-navbar.component.html', selector: 'header-navbar', styleUrls: ['./navbar.component.css'] })
+
+export class NavbarComponent implements OnInit {
+
+    ngOnInit() {
+    }
+      constructor(private modalService: ModalService) {}
+
+      onOpenModal(mode: string) {
+      this.modalService.onOpenModal(mode)
+    }
+}
