+++ employee-manager-app/src/app/components/user-employee/user.component.html
@@ -1,10 +1,4 @@
-<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <a class="navbar-brand" style="color: white">Employ Easy</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarColor02"
-        aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-    </button>
-</nav>
+<header-navbar></header-navbar> 
 <div class="container" id="main-container">
     <h5 class="main-hello">Hello {{ employee?.firstName }} !</h5>
     <div class="row justify-content-center align-items-center">
@@ -34,7 +28,6 @@
                     <li class="list-group-item"><i class="fa fa-envelope float-right"></i>Personal Number: {{ employee?.personalNumber }}</li>
                     <li class="list-group-item"><i class="fa fa-envelope float-right"></i>Start Date: {{ employee?.startDate }}</li>
                     <li class="list-group-item"><i class="fa fa-envelope float-right"></i>End Date: {{ employee?.endDate }}</li>
-
                     <li class="list-group-item">
                       <!---  <i class="fa fa-phoneNumber float-right"></i>Phone : {{ employee?.phoneNumber }} -->
                     </li>
