+++ employee-manager-app/src/app/services/login.service.ts
@@ -8,7 +8,6 @@ import { environment } from 'src/environments/environment';
 @Injectable({ providedIn: 'root' })
 export class LoginService {
 
-
     constructor(
         private http: HttpClient
     ) {}
